fixed trips migration foreign key

+++ database/migrations/2023_05_19_223919_create_trips_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('trips', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('car_id');
+            $table->unsignedBigInteger('car_id');
             $table->date('date');
             $table->integer('miles');
             $table->timestamps();
